fix dpcpp

+++ src/shamalgs/src/primitives/is_all_true.cpp
@@ -76,7 +76,7 @@ namespace {
         auto dev_sched = buf.get_dev_scheduler_ptr();
         auto &q        = dev_sched->get_queue();
 
-        sham::DeviceBuffer<u8> stop_flag(1, dev_sched);
+        sham::DeviceBuffer<u32> stop_flag(1, dev_sched);
         stop_flag.fill(0);
 
         /*
@@ -126,7 +126,7 @@ namespace {
                         // Only the group leader reads the stop flag from device memory,
                         // then broadcast that single value to the rest of the group instead
                         // of every work-item issuing its own global memory load.
-                        u8 stop_val = sycl::group_broadcast(grp, (lid == 0) ? *stop : u8{0}, 0);
+                        u32 stop_val = sycl::group_broadcast(grp, (lid == 0) ? *stop : u32{0}, 0);
 
                         // early exit the whole group if the flag is set
                         if (stop_val) {
@@ -144,12 +144,12 @@ namespace {
                             // if there is a false we set the stop flag
                             if (!result && !(*stop)) {
                                 sycl::atomic_ref<
-                                    u8,
+                                    u32,
                                     sycl::memory_order_relaxed,
                                     sycl::memory_scope_device,
                                     sycl::access::address_space::global_space>
                                     atom(*stop);
-                                atom |= 1_u8;
+                                atom |= 1_u32;
                             }
                         }
                     });
